Add files via upload
Learnt the Arrays and a bit of Vectors. In 12th grade, my definition of Vector was different but when I am learning C++, it feels like coming to another dimension especially LEARNING how flexible Vector is against Arrays gave me a new way of thinking. Vectors actually helped in cutting the excessive length of my codes which I am grateful of .